Fixed last commit.
Last commit attempted to patch out memory leaks in the case memory allocated on the heap was NULL and dereferenced.

However, there were badly designed for loops which permently shifted the address spaces of pointers which had to be reprimanded.

All is good now, and there are no memory leaks.

Hooray! o,,o